requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
dspy-ai>=2.0.0
anthropic>=0.5.0
requests>=2.28.0 